+++ mycodo/mycodo_flask/routes_page.py
@@ -1551,6 +1551,7 @@ def page_data():
     form_mod_math = forms_math.MathMod()
     form_mod_math_measurement = forms_math.MathMeasurementMod()
     form_mod_average_single = forms_math.MathModAverageSingle()
+    form_mod_redundancy = forms_math.MathModRedundancy()
     form_mod_difference = forms_math.MathModDifference()
     form_mod_equation = forms_math.MathModEquation()
     form_mod_humidity = forms_math.MathModHumidity()
@@ -1619,6 +1620,8 @@ def page_data():
                 utils_math.math_mod(form_mod_math, form_mod_humidity)
             elif math_type == 'average_single':
                 utils_math.math_mod(form_mod_math, form_mod_average_single)
+            elif math_type == 'redundancy':
+                utils_math.math_mod(form_mod_math, form_mod_redundancy)
             elif math_type == 'difference':
                 utils_math.math_mod(form_mod_math, form_mod_difference)
             elif math_type == 'equation':
@@ -1733,11 +1736,12 @@ def page_data():
                            form_mod_input=form_mod_input,
                            form_mod_input_measurement=form_mod_input_measurement,
                            form_mod_average_single=form_mod_average_single,
-                           form_mod_math=form_mod_math,
-                           form_mod_math_measurement=form_mod_math_measurement,
+                           form_mod_redundancy=form_mod_redundancy,
                            form_mod_difference=form_mod_difference,
                            form_mod_equation=form_mod_equation,
                            form_mod_humidity=form_mod_humidity,
+                           form_mod_math=form_mod_math,
+                           form_mod_math_measurement=form_mod_math_measurement,
                            form_mod_verification=form_mod_verification,
                            form_mod_misc=form_mod_misc,
                            input_templates=input_templates,