+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
  - Add Additional check for Conditional Statements if {ID} is replaced with None ([#571](https://github.com/kizniche/mycodo/issues/571))
  - Add ability to set Logging Interval and download logged data from SHT31 Smart Gadget ([#559](https://github.com/kizniche/mycodo/issues/559))
+ - Add Math: Input Backup: If a measurement of an Input cannot be found, look for a measurement of another (or another, etc.) ([#559](https://github.com/kizniche/mycodo/issues/559))
 
 ### Miscellaneous